+++ src/core/cache/decorators/ttl/interface.ts
@@ -14,6 +14,11 @@ export interface TTLCache<
 	T extends CacheWithEmitter<K, V> = CacheWithEmitter<K, V>
 > extends CacheWithEmitter<K, V> {
 	/**
+	 * Collection of cache's timers
+	 */
+	ttlTimers: Map<K, [number | NodeJS.Timeout, Promise<CanUndef<V>>]>;
+
+	/**
 	 * Saves a value to the cache by the specified key
 	 *
 	 * @param key